models/model/requirements.txt
numpy==1.20.3
pillow==8.2.0
torch==1.9.1+cpu
pandas==1.3.3
torchvision==0.9.0
